requirements.txt
Django==4.1
djangorestframework==3.13.1
requests~=2.28.1
beautifulsoup4~=4.11.1
python-dotenv==0.19.2
django-filter~=22.1